requirements.txt
Flask==3.0.0
gunicorn==21.2.0
bcrypt==4.1.3
requests==2.31.0
python-dotenv==1.0.0
cffi>=1.15.0
passlib==1.7.4

# Google Cloud integrations
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.8.0

